backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-json-logger==2.0.7
openai==1.51.2
faiss-cpu==1.9.0
numpy==2.1.2
httpx==0.27.2
